+++ config-presets/modern/vite.config.ts
@@ -1,32 +1,20 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
-import legacy from '@vitejs/plugin-legacy'
 
 // https://vite.dev/config/
 export default defineConfig({
+  base: './',
   plugins: [
     react(),
-    tailwindcss(),
-    legacy({
-      targets: ['safari >= 13', 'ios >= 13', 'chrome >= 71', 'edge >= 79', 'firefox >= 67', 'not IE 11'],
-      additionalLegacyPolyfills: ['regenerator-runtime/runtime'],
-      polyfills: [
-        'es.promise.finally', 
-        'es.array.flat-map', 
-        'es.array.flat', 
-        'es.object.from-entries',
-        'es.symbol.description'
-      ],
-      modernPolyfills: true
-    })
+    tailwindcss()
   ],
   build: {
-    target: 'es2015',
+    target: 'es2020',
     minify: 'terser',
     terserOptions: {
       safari10: true,
     },
-    cssTarget: 'safari13'
+    cssTarget: 'safari15'
   }
 })